refactor(analysis): split loop state/guards out of _loops
